Trying to improve the color palette
I set up a color palette in the README, and then changed the header,
footer, and sum/total to reflect them.

I also moved the sum/total above the score adjusters so that it mimics
a calculator better.

The width is now 100%, so that it does not have a wide margin when using
a desktop.